backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
torch==2.7.0
torchvision==0.22.0
sentence-transformers==2.7.0
transformers==4.35.2
Pillow==10.1.0
requests==2.31.0
pinecone-client==3.0.0
langchain==0.0.350
